Support up to INT_MAX (2^31-1) for the absolute value of css <number> data type.
When converting parsed number from string to double do not use a flot variable to store
the result as it does not have the necessary precision to represent exactly all the values
in the 32-bit integer range. Use a double instead.

BUG: 335346
FIXED-IN: 4.13.2

+++ khtml/css/parser.cpp
@@ -289,7 +289,7 @@ typedef union YYSTYPE
     CSSMediaRuleImpl *mediaRule;
     CSSRuleListImpl *ruleList;
     ParseString string;
-    float val;
+    double val;
     int prop_id;
     unsigned int attribute;
     unsigned int element;
